Remove duplicated functions, move unit tests into the right location
